refactor(telegram): move the html surgery into TelegramHtml and unify prose entries
